+++ services/v1/manifest.js
@@ -15,6 +15,7 @@ const rootPrefix = "../.."
   , transfersKlass = require(rootPrefix + '/services/v1/transfers')
   , usersKlass = require(rootPrefix + '/services/v1/users')
   , balancesKlass = require(rootPrefix + '/services/v1/balances')
+  , ledgerKlass = require(rootPrefix + '/services/v1/ledger')
 ;
 
 // hide request object
@@ -39,6 +40,7 @@ const manifest = function (params) {
   oThis.transfers = new transfersKlass(_requestObj);
   oThis.users = new usersKlass(_requestObj);
   oThis.balances = new balancesKlass(_requestObj);
+  oThis.ledger = new ledgerKlass(_requestObj);
 
   return oThis;
 };
@@ -64,7 +66,10 @@ manifest.prototype = {
   users: null,
 
   // Services at /balances endpoint
-  balances: null
+  balances: null,
+
+  // Services at /ledger endpoint
+  ledger: null
 
 };