+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[OST Javascript SDK v1.0.1](https://github.com/OpenSTFoundation/ost-sdk-js/tree/v1.0.1) June 20 2018
+---
+
+* Added balance module to v1 API
+* Added ledger module to v1 API
+
 [OST Javascript SDK v1.0.0](https://github.com/OpenSTFoundation/ost-sdk-js/tree/v1.0.0) May 17 2018
 ---